Fixed compile errors for encode Base64 with FeebeeCam. Captures well and with reasonable time

+++ server/b-string/test.h
@@ -90,7 +90,7 @@ namespace BeeFishBString
       ok &= testResult(
          "Write bit",
          (data.size() == 1 &&
-          data[0] == 0b10000000)
+          data[0] == (Byte)0b10000000)
       );
       
       BitStream readStream =
